move pytest to its own task file
